# 4.07.23(2)/4.07.23(2)/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "4.07.23(2)",
    "version": "1.0"
  },
  "paths": {
...
    "/api/IsBakanMeslek": {
      "get": {
        "tags": [
          "IsBakanMeslek"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "IsBakanMeslek"
        ],
        "parameters": [
          {
            "name": "Jsid",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Meslekid",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "IsBakanMeslek"
        ],
        "parameters": [
          {
            "name": "JsId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "MeslekId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meslek"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Meslek"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Meslek"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/IsBakanMeslek/{id}": {
      "delete": {
        "tags": [
          "IsBakanMeslek"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/IsSahibiAPI": {
      "get": {
        "tags": [
          "IsSahibiAPI"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IsSahibi"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IsSahibi"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IsSahibi"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "IsSahibiAPI"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IsSahibi"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsSahibi"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsSahibi"
                }
              }
            }
          }
        }
      }
    },
    "/api/IsSahibiAPI/{id}": {
      "get": {
        "tags": [
          "IsSahibiAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IsSahibi"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsSahibi"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsSahibi"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "IsSahibiAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "IsSahibiAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/JobSeekerAPI/{mail}": {
      "get": {
        "tags": [
          "JobSeekerAPI"
        ],
        "parameters": [
          {
            "name": "mail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "JobSeekerAPI"
        ],
        "parameters": [
          {
            "name": "mail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Meslek/{id}": {
      "get": {
        "tags": [
          "Meslek"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Meslek"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Meslek"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IsSahibi": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          },
          "soyisim": {
            "type": "string",
            "nullable": true
          },
          "mail": {
            "type": "string",
            "nullable": true
          },
          "telno": {
            "type": "integer",
            "format": "int64"
          },
          "kullanıcıSifresi": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Meslek": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}